Reorganize Redux folder
