+++ src/components/AddApplication/AddApplication.js
@@ -3,7 +3,7 @@ import { useHistory } from 'react-router-dom';
 import { useSelector, useDispatch, shallowEqual } from 'react-redux';
 import { useIntl } from 'react-intl';
 
-import { loadEntities } from '../../redux/actions/sources';
+import { loadEntities } from '../../redux/sources/actions';
 import SourcesFormRenderer from '../../Utilities/SourcesFormRenderer';
 import createSchema from './AddApplicationSchema';
 import LoadingStep from '../steps/LoadingStep';
