+++ src/components/RedirectNoId/RedirectNoId.js
@@ -3,7 +3,7 @@ import { useDispatch, useSelector, shallowEqual } from 'react-redux';
 import { useParams, Redirect } from 'react-router-dom';
 import { useIntl } from 'react-intl';
 
-import { addMessage, addHiddenSource } from '../../redux/actions/sources';
+import { addMessage, addHiddenSource } from '../../redux/sources/actions';
 import { doLoadSource } from '../../api/entities';
 import { useIsLoaded } from '../../hooks/useIsLoaded';
