+++ src/components/UndoButton/UndoButtonAdd.js
@@ -5,7 +5,7 @@ import { useSelector, useDispatch } from 'react-redux';
 import { useHistory } from 'react-router-dom';
 import { FormattedMessage } from 'react-intl';
 
-import { removeMessage, undoAddSource } from '../../redux/actions/sources';
+import { removeMessage, undoAddSource } from '../../redux/sources/actions';
 import { paths } from '../../Routes';
 import { refreshPage } from './refreshPage';
