+++ src/redux/sources/actions.js
@@ -10,7 +10,7 @@ import {
     SET_COUNT,
     ADD_HIDDEN_SOURCE,
     CLEAR_FILTERS
-} from '../action-types-sources';
+} from './actions-types';
 import {
     doLoadAppTypes,
     doRemoveSource,
